refactor: update error imports and add MatchFieldValidation in ValidationBuilder test case
