+++ src/validation/errors/index.ts
@@ -1,2 +1,3 @@
 export * from './RequiredFieldError';
 export * from './InvalidFieldError';
+export * from './MatchFieldError'
